refactor: BaseDto 위치 변경
